not quite all overflow errors were fixed
svn path=/branches/SLED10-SP1/; revision=264

+++ libslab/double-click-detector.h
@@ -37,7 +37,7 @@ typedef struct
 	GObject parent_placeholder;
 
 	guint32 double_click_time;
-	guint32 last_click_time;
+	guint64 last_click_time;
 } DoubleClickDetector;
 
 typedef struct
@@ -49,10 +49,11 @@ GType double_click_detector_get_type (void);
 
 DoubleClickDetector *double_click_detector_new (void);
 
-gboolean double_click_detector_is_double_click (DoubleClickDetector * detector, guint32 event_time,
+gboolean double_click_detector_is_double_click (DoubleClickDetector * detector, guint64 event_time,
 	gboolean auto_update);
 
-void double_click_detector_update_click_time (DoubleClickDetector * detector, guint32 event_time);
+void double_click_detector_update_click_time (DoubleClickDetector * detector, guint64 event_time);
 
 G_END_DECLS
+
 #endif /* __DOUBLE_CLICK_DETECTOR_H__ */
